+++ Modules/ActiveDirectoryDsc.Common/ActiveDirectoryDsc.Common.psm1
@@ -1831,6 +1831,7 @@ function Test-DscPropertyState
             $supportedTypes = @(
                 'String'
                 'Int32'
+                'UInt32'
                 'Int16'
                 'UInt16'
                 'Single'